refactor(agent): 迁移模型解析到 workflow/resolver 子包
